Fix for the moment
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@135655 c90b9560-bf6c-de11-be94-00142212c4b1

+++ Zend/zend_hash.h
@@ -288,7 +288,7 @@ END_EXTERN_C()
 
 static inline int zend_symtable_update(HashTable *ht, char *arKey, uint nKeyLength, void *pData, uint nDataSize, void **pDest)					\
 {
-	int idx;
+	long idx;
 
 	if (zend_is_numeric_key(arKey, nKeyLength, &idx)) {
 		return zend_hash_index_update(ht, idx, pData, nDataSize, pDest);
@@ -300,7 +300,7 @@ static inline int zend_symtable_update(HashTable *ht, char *arKey, uint nKeyLeng
 
 static inline int zend_symtable_del(HashTable *ht, char *arKey, uint nKeyLength)
 {
-	int idx;
+	long idx;
 
 	if (zend_is_numeric_key(arKey, nKeyLength, &idx)) {
 		return zend_hash_index_del(ht, idx);
@@ -312,7 +312,7 @@ static inline int zend_symtable_del(HashTable *ht, char *arKey, uint nKeyLength)
 
 static inline int zend_symtable_find(HashTable *ht, char *arKey, uint nKeyLength, void **pData)
 {
-	int idx;
+	long idx;
 
 	if (zend_is_numeric_key(arKey, nKeyLength, &idx)) {
 		return zend_hash_index_find(ht, idx, pData);
@@ -324,7 +324,7 @@ static inline int zend_symtable_find(HashTable *ht, char *arKey, uint nKeyLength
 
 static inline int zend_symtable_exists(HashTable *ht, char *arKey, uint nKeyLength)
 {
-	int idx;
+	long idx;
 
 	if (zend_is_numeric_key(arKey, nKeyLength, &idx)) {
 		return zend_hash_index_exists(ht, idx);
